+++ development/calendar/main.tsx
@@ -78,6 +78,7 @@ const calendar = createCalendar({
   //   nEventsPerDay: 7
   // },
   firstDayOfWeek: 1,
+  // locale: 'de-DE',
   // locale: 'ja-JP',
   // locale: 'en-US',
   // locale: 'zh-CN',
@@ -89,7 +90,7 @@ const calendar = createCalendar({
   // defaultView: viewWeek.name,
   // minDate: '2024-01-01',
   // maxDate: '2024-03-31',
-  defaultView: 'week',
+  // defaultView: 'week',
   // selectedDate: '2024-12-01',
   // datePicker: {
   //   selectedDate: '2023-11-01'