+++ packages/shared/src/interfaces/calendar/calendar-config.ts
@@ -22,7 +22,6 @@ export type WeekOptions = {
   nDays: number
   eventWidth: number
   timeAxisFormatOptions: Intl.DateTimeFormatOptions
-  showLocation: boolean
 }
 
 export type MonthGridOptions = {